requirements.txt
pytorch-lightning
torchvision
albumentations
numpy
wandb
hydra-core